% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rarefaction_curve.R
\name{amp_rarecurve}
\alias{amp_rarecurve}
\title{Generate Rarefaction Curves}
\usage{
amp_rarecurve(
  data,
  step = 100,
  ylim = NULL,
  xlim = NULL,
  label = FALSE,
  color = NULL,
  rank = rank,
  legend = TRUE,
  color.vector = NULL,
  legend.position = "topleft"
)
}
\arguments{
\item{data}{A phyloseq object containing OTU data and associated metadata.}

\item{step}{An integer specifying the step size for the rarefaction curve. Default is 100.}

\item{ylim}{A numeric vector of length two specifying the y-axis limits. Default is \code{NULL},
which automatically adjusts the y-axis.}

\item{xlim}{A numeric vector of length two specifying the x-axis limits. Default is \code{NULL},
which automatically adjusts the x-axis.}

\item{label}{A logical value indicating whether sample labels should be displayed on the
plot. Default is \code{FALSE}.}

\item{color}{A character string specifying the column in the sample metadata to use for
coloring the samples. Default is \code{NULL} (no color applied).}

\item{rank}{An optional rank level for taxa aggregation. Default is \code{rank}.}

\item{legend}{A logical value indicating whether to display a legend on the plot.
Default is \code{TRUE}.}

\item{color.vector}{An optional character vector specifying custom colors for groups in
the \code{color} column. Default is \code{NULL} (default colors are used).}

\item{legend.position}{A character string specifying the position of the legend. Default
is \code{"topleft"}.}
}
\value{
A rarefaction curve plot.
}
\description{
This function generates rarefaction curves for visualizing species richness
at different sequencing depths, based on the OTU table in a phyloseq object.
}
\examples{
amp_rarecurve(data = physeq, step = 200, color = "sample_type", legend.position = "topright")
}
