% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_folders.R
\name{create_folders}
\alias{create_folders}
\title{Create Project Folder Structure}
\usage{
create_folders(projects)
}
\arguments{
\item{projects}{A character vector containing the name of the project folders.
Each project folder must already exist within the specified
base path. The \code{projects} argument represents the name of the
project folder, which must contain a subfolder named \code{qiime2_output}
with the output files generated from QIIME2. This folder structure
is necessary for setting up downstream analysis.
\code{projects} is typically reference in a configuration file.}

\item{base_path}{A character string indicating the base directory where
project folder is located. This path is used to locate
the project folder and create the required subfolders.
\code{base_path} is typically referenced in a configuration file.}
}
\value{
None. This function is called for its side effects.
}
\description{
This function creates a folder structure for projects, ensuring that
all necessary directories exist and that specific files required for downstream
analysis are copied from the \code{qiime2_output} folder into the \code{input_data} folder.
}
\details{
This function is designed to facilitate the setup of downstream analyses by:
\itemize{
\item Creating a consistent directory structure for each project, including
subfolders like \code{input_data}, \code{output_data}, \code{figures}, and \code{messages}.
\item Copying essential files from the \code{qiime2_output} folder into the \code{input_data} folder.
These essential files include:
\itemize{
\item \code{table.qza}: The feature table from QIIME2.
\item \code{rooted-tree.qza}: The phylogenetic tree used for diversity analysis.
\item \code{classifier.qza}: The classifier used for taxonomy assignment.
\item \code{metadata.tsv}: The QIIME2 sample metadata file required for analyses.
\item \code{dna-sequences.csv}: Contains 16S rRNA gene copy number predictions.
\item \code{metadata_extra.tsv}: Any additional metadata provided for the samples.
}
\item Checking for optional files that enhance analyses, such as:
\itemize{
\item \code{qPCR.csv}: Contains quantitative PCR data.
\item \code{fcm.csv}: Contains flow cytometry data.
}
\item Logging warnings for missing optional files and errors for missing required files.
}

The function ensures that downstream analyses, which rely on specific input files
(e.g., the \code{table.qza}, \code{rooted-tree.qza}, etc.), have access to these files in the
correct directory structure. If any required files are missing from the
\code{qiime2_output} folder, the function stops execution and logs an error message.
}
\note{
Each project folder must already exist within the \code{base_path} directory and must
contain a subfolder named \code{qiime2_output}, which holds the outputs of QIIME2 analysis.
The function sets up the folder structure for downstream analysis within this project folder.
}
\examples{
\dontrun{
# Define the base path and log file location
base_path <- "path/to/projects"
log_file <- "path/to/logfile.txt"

# Create folder structures for two projects
create_folders(projects)
}

}
