% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/heatmap.R
\name{heatmap}
\alias{heatmap}
\title{Generate a Heatmap of Relative Abundance}
\usage{
heatmap(
  psdata,
  taxrank,
  ntaxa = 23,
  facet_vars,
  plot_width,
  project_id,
  base_path,
  log_file
)
}
\arguments{
\item{taxrank}{A character string indicating the taxonomic rank to use for grouping taxa. The default is \code{"Genus"}.}

\item{ntaxa}{An integer specifying the maximum number of taxa to display individually. Taxa below the threshold are grouped into "Other". If \code{NULL}, \code{ntaxa} is set to 23.}

\item{physeq}{A phyloseq object containing normalized genus-level data. The default is \code{rarefied_genus_psmelt}.}

\item{norm_method}{A character string specifying the normalization method. If \code{NULL}, the function uses the provided \code{physeq} directly. If set to \code{"fcm"} or \code{"qpcr"}, the function extracts the corresponding \code{psmelt_copy_number_corrected_} data based on the taxonomic rank.}
}
\value{
A \code{ggplot} object representing the heatmap of relative abundance.
}
\description{
This function creates a heatmap of relative abundance data from a phyloseq
object (or similar data frame) at the genus level. It calculates the relative
abundance of each taxon per sample and groups taxa with low relative
abundance (below a defined threshold) into an "Other" category. The heatmap
is then facetted based on additional sample metadata if available and saved
as a PDF.
}
\details{
The function performs the following steps:
\enumerate{
\item Sets up project folder paths for figures and output data.
\item Extracts and processes the input data to compute the relative abundance (in percentage) of each taxon per sample.
\item Groups taxa with a mean relative abundance below a defined cutoff into an "Other" category.
\item Optionally orders the data by \code{Sample_Date} if that factor is present in the metadata.
\item Creates a base heatmap using \code{ggplot2}, with samples on the x-axis and taxa on the y-axis. The fill color reflects the relative abundance, and text labels are added for values exceeding a threshold.
\item If more than one \code{na_type} is present (e.g., both DNA and RNA), separate heatmaps are generated for each and then combined.
\item Saves the final heatmap as a PDF file in the project's figures folder.
}
}
\examples{
\dontrun{
  # Generate a heatmap using default parameters
  heatmap_plot <- heatmap(physeq = rarefied_genus_psmelt)

  # Generate a heatmap with a specified number of taxa and a normalization method
  heatmap_plot <- heatmap(
    physeq = rarefied_genus_psmelt,
    ntaxa = 20,
    norm_method = "fcm",
    taxrank = c("Phylum", "Class", "Order", "Family", "Genus")
  )
}

}
