% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hello.R
\name{hello}
\alias{hello}
\title{Hello, world!}
\usage{
hello()
}
\value{
Geen waarde, alleen een print naar de console.
}
\description{
Dit is een voorbeeldfunctie genaamd 'hello', die 'Hello, world!' afdrukt.
}
\examples{
hello()

}
