% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/physeq_to_taxa_tibbles.R
\name{physeq_to_taxa_tibbles}
\alias{physeq_to_taxa_tibbles}
\title{Aggregate Phyloseq Data by Taxonomic Level}
\usage{
physeq_to_taxa_tibbles(
  physeq,
  norm_method = NULL,
  copy_correction = TRUE,
  taxrank = c("Phylum", "Class", "Order", "Family", "Genus"),
  project_id,
  base_path,
  log_file
)
}
\arguments{
\item{taxrank}{A character vector indicating the taxonomic levels at which to group the data.}
}
\value{
The function saves multiple \code{phyloseq} objects as RDS files.
The aggregated objects are saved in the output directory \verb{output_data/rds_files/After_cleaning_rds_files/}.
}
\description{
This function aggregates ASV data at specified taxonomic levels (e.g., Phylum, Class, Order, Family, or Genus)
using the \code{tax_glom} function from the \pkg{phyloseq} package.
}
\details{
The function applies the \code{tax_glom} function to group ASVs at each specified taxonomic level. It creates
a dedicated folder for each taxonomic level under the output directory and saves the aggregated data as RDS files.
}
\examples{
\dontrun{
# Aggregate data using flow cytometry normalization
result <- group_tax(physeq = rarefied_asv_physeq, norm_method = "fcm")

# Aggregate data using qPCR normalization
result <- group_tax(physeq = rarefied_asv_physeq, norm_method = "qpcr")
}

}
