% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rarefaction_curve.R
\name{rarefaction_curve}
\alias{rarefaction_curve}
\title{Generate and Save Rarefaction Curve}
\usage{
rarefaction_curve(
  physeq,
  color = "sample_or_control",
  project_id,
  base_path,
  log_file
)
}
\arguments{
\item{physeq}{A phyloseq object to be analyzed.}

\item{color}{Character string specifying the metadata column used for coloring lines.
Defaults to \code{"sample_or_control"}.}

\item{project_id}{Character string specifying the unique project name.}

\item{base_path}{Character string specifying the root project directory.}

\item{log_file}{Character string specifying the path to the log file.}
}
\value{
None. This function is called for its side effects (saving a PDF file).
}
\description{
Generates a sequencing depth rarefaction curve from a phyloseq object
and exports the resulting plot as a PDF.
}
