% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rarefy_counts.R
\name{rarefy_counts}
\alias{rarefy_counts}
\title{Rarefy counts}
\usage{
rarefy_counts(
  physeq,
  norm_method = NULL,
  copy_correction = TRUE,
  iteration = 100,
  project_id,
  base_path,
  log_file
)
}
\arguments{
\item{physeq}{A \code{phyloseq} object containing the microbiome data.
This is the input object that the function processes.}

\item{norm_method}{A character string specifying the normalization method. Acceptable values are:
\itemize{
\item \code{NULL}: Use this option if no FCM or qPCR data is available, or if you wish to retain only relative abundances.
\item \code{"fcm"}: Use this option if the data have been normalized using flow cytometry (FCM).
\item \code{"qpcr"}: Use this option if the data have been normalized using quantitative PCR (qPCR).
}}

\item{copy_correction}{A logical value indicating whether the data should be corrected for
the predicted 16S rRNA copy numbers prior to biomass normalization. Options are:
\itemize{
\item \code{TRUE}: Both relative and absolute abundances are corrected using the predicted copy numbers.
\item \code{FALSE}: Abundances are not corrected by copy number. Note that qPCR normalization requires
copy number correction to provide absolute data.
}}
}
\value{
The rarefied \code{phyloseq} object is returned and also saved as an \code{.rds} file. The file name and location
depend on the specified normalization method:
\itemize{
\item For \code{"fcm"}: \code{"project_name_phyloseq_asv_level_fcm_normalised_cell_concentration_rarefied.rds"}
\item For \code{"qpcr"}: \code{"project_name_phyloseq_asv_level_qpcr_normalised_cell_concentration_rarefied.rds"}
}
}
\description{
This function performs rarefaction on a normalized \code{phyloseq} object.
Rarefaction is based on the biomass of each sample to identify the minimum
sampling depth across the dataset. By rarefying all samples to this sampling
depth, we ensure that the data are normalized for sequencing effort. This
step prevents an overestimation of genus abundance in samples with higher
sequencing depths, allowing for more accurate and comparable results.
}
\details{
\itemize{
\item For \code{"fcm"} normalization:
\itemize{
\item Rarefies only the \code{fcm} normalized data, while the \verb{copy number corrected} data remains unchanged.
\item Rarefies based on the calculated sampling depth, derived from the ratio of total reads per sample to the
estimated cell counts.
\item Uses a custom function (\code{avgrarefy}) to perform multiple iterations of rarefaction and averages the results.
}
\item For \code{"qpcr"} normalization:
\itemize{
\item Rarefies only the \code{qpcr} normalized data, while the \verb{copy number corrected} data remains unchanged.
\item Rarefies based on the calculated sampling depth, derived from the ratio of total reads per sample to
predicted 16S rRNA gene copy numbers.
\item Uses the same \code{avgrarefy} function for averaging rarefied counts.
}
\item The input values are subject to a scaling limit of 1e7. If input values exceed this limit due to the prior normalization,
all values are scaled down by a calculated scaling factor. The rarefied counts are rescaled back to their original scale
after rarefaction. However, due to this scaling and the rounding of scaled values, slight variations in the rarefied counts
may occur.
}

The function uses parallel processing to improve the efficiency of rarefaction. It saves the rarefied \code{phyloseq}
object as an \code{.rds} file in the appropriate output folder.
}
\note{
Ensure that the \code{phyloseq} object is properly normalized before applying this function.
Missing or invalid \code{rarefy_to} values will result in warnings and skipped samples.
}
\examples{
# Rarefy using FCM normalization
rarefied_physeq <- rarefy_counts(physeq = normalised_physeq, norm_method = "fcm")

# Rarefy using qPCR normalization
rarefied_physeq <- rarefy_counts(physeq = normalised_physeq, norm_method = "qpcr")

}
