% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rarefying.R
\name{rarefying}
\alias{rarefying}
\title{Rarefying Phyloseq Data}
\usage{
rarefying(physeq = physeq, norm_method = NULL, iteration = 100)
}
\arguments{
\item{physeq}{A list containing the \code{phyloseq} objects. Required components:
\itemize{
\item \code{psdata_asv_copy_number_corrected}: With copy number corrected phyloseq data.
\item \code{psdata_asv_fcm_norm}: Required if \code{norm_method} = "fcm".
\item \code{psdata_asv_qpcr_norm}: Required if \code{norm_method} = "qpcr".
}}

\item{norm_method}{A string specifying the normalization method. Acceptable values:
\itemize{
\item \code{NULL}: Returns the \code{psdata_asv_copy_number_corrected} object without modifications.
\item \code{"fcm"}: Rarefied data using FCM normalized.
\item \code{"qpcr"}: Rarefied data using qPCR normalized.
}}
}
\value{
The rarefied \code{phyloseq} object is returned and also saved as an \code{.rds} file. The file name and location
depend on the specified normalization method:
\itemize{
\item For \code{"fcm"}: \code{"project_name_phyloseq_asv_level_fcm_normalised_cell_concentration_rarefied.rds"}
\item For \code{"qpcr"}: \code{"project_name_phyloseq_asv_level_qpcr_normalised_cell_concentration_rarefied.rds"}
}
}
\description{
This function performs rarefaction on a normalized \code{phyloseq} object.
Rarefaction is based on the biomass of each sample to identify the minimum
sampling depth across the dataset. By rarefying all samples to this sampling
depth, we ensure that the data are normalized for sequencing effort. This
step prevents an overestimation of genus abundance in samples with higher
sequencing depths, allowing for more accurate and comparable results.
}
\details{
\itemize{
\item For \code{"fcm"} normalization:
\itemize{
\item Rarefies only the \code{fcm} normalized data, while the \verb{copy number corrected} data remains unchanged.
\item Rarefies based on the calculated sampling depth, derived from the ratio of total reads per sample to the
estimated cell counts.
\item Uses a custom function (\code{avgrarefy}) to perform multiple iterations of rarefaction and averages the results.
}
\item For \code{"qpcr"} normalization:
\itemize{
\item Rarefies only the \code{qpcr} normalized data, while the \verb{copy number corrected} data remains unchanged.
\item Rarefies based on the calculated sampling depth, derived from the ratio of total reads per sample to
predicted 16S rRNA gene copy numbers.
\item Uses the same \code{avgrarefy} function for averaging rarefied counts.
}
\item The input values are subject to a scaling limit of 1e7. If input values exceed this limit due to the prior normalization,
all values are scaled down by a calculated scaling factor. The rarefied counts are rescaled back to their original scale
after rarefaction. However, due to this scaling and the rounding of scaled values, slight variations in the rarefied counts
may occur.
}

The function uses parallel processing to improve the efficiency of rarefaction. It saves the rarefied \code{phyloseq}
object as an \code{.rds} file in the appropriate output folder.
}
\note{
Ensure that the \code{phyloseq} object is properly normalized before applying this function.
Missing or invalid \code{rarefy_to} values will result in warnings and skipped samples.
}
\examples{
# Rarefy using FCM normalization
rarefied_physeq <- rarefying(physeq = normalised_physeq, norm_method = "fcm")

# Rarefy using qPCR normalization
rarefied_physeq <- rarefying(physeq = normalised_physeq, norm_method = "qpcr")

}
