% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/alpha_diversity.R
\name{alpha_diversity}
\alias{alpha_diversity}
\title{Generate Alpha Diversity Plots}
\usage{
alpha_diversity(
  physeq = physeq,
  norm_method = NULL,
  taxrank = c("Phylum", "Class", "Order", "Family", "Tax_label"),
  date_factor = NULL
)
}
\arguments{
\item{physeq}{A phyloseq object containing microbial community data.}

\item{norm_method}{A character string specifying the normalization method for the data. Options include:
\itemize{
\item \code{"fcm"}: Use flow cytometry-normalized data.
\item \code{"qpcr"}: Use qPCR-normalized data.
\item \code{NULL}: Use only copy number corrected data (default).
}}

\item{taxrank}{A character vector specifying the taxonomic levels for which
alpha diversity is to be calculated. If the first element is (taxrank = \code{asv}),
ASV-level data is processed. Otherwise, the function
processes data for each taxonomic level provided default is taxrank = c('Phylum', 'Class', 'Order', 'Family', 'Tax_label').}

\item{date_factor}{An optional character string indicating the name of the
date column in the sample metadata. If provided, the column is converted to
a Date object ("\%d/\%m/\%Y") and used to order the data.}
}
\value{
A combined ggplot object containing the generated alpha diversity plots.
}
\description{
This function calculates alpha diversity metrics from a phyloseq object and
generates alpha diversity plots at various taxonomic levels or at the ASV
level. The alpha diversity measures (Observed, Chao1, Shannon, Simpson) are
computed using the \code{estimate_richness} function from the phyloseq
package. Depending on the input parameters, the function can handle
normalized data (using flow cytometry or qPCR methods) and can separate plots
by DNA and RNA types.
}
\details{
The function performs the following steps:
\enumerate{
\item Extracts the appropriate data object from \code{physeq} based on the chosen \code{norm_method} and taxonomic level.
\item Estimates alpha diversity metrics (Observed, Chao1, Shannon, Simpson) using \code{estimate_richness}.
\item Merges the alpha diversity estimates with sample metadata.
\item Optionally orders the data by a date factor if provided.
\item Appends dummy data to the dataset for visualization purposes.
\item Exports the combined alpha diversity data as CSV files.
\item Generates bar plots for the diversity metrics, optionally separating DNA and RNA data if both are present.
\item Saves the resulting plots as PDF files in the project's figures folder.
}
}
\examples{
\dontrun{
  # Generate alpha diversity plots at the ASV level without normalization
  alpha_plot <- alpha_diversity(physeq = my_physeq, taxrank = "asv")

  # Generate alpha diversity plots at the Phylum level using flow cytometry-normalized data
  alpha_plot <- alpha_diversity(physeq = my_physeq, norm_method = "fcm", taxrank = "Phylum")

  # Generate alpha diversity plots with a specified date factor for ordering samples
  alpha_plot <- alpha_diversity(physeq = my_physeq, date_factor = "Sample_Date")
}

}
