% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/barplot.R
\name{barplot}
\alias{barplot}
\title{Create Barplots for Relative and Absolute Abundance}
\usage{
barplot(
  physeq,
  ntaxa = NULL,
  norm_method = NULL,
  taxrank = c("Phylum", "Class", "Order", "Family", "Genus")
)
}
\arguments{
\item{ntaxa}{An integer specifying the maximum number of taxa to display in the barplot. Default is 23.}

\item{norm_method}{A string indicating the normalization method used for absolute abundance
data. Options are \code{"fcm"} (flow cytometry) or \code{"qpcr"} (quantitative PCR).
(relative abundance only).}

\item{sample_matrix}{An optional matrix specifying the sample structure or metadata.}

\item{group_by_factor}{with this option you can separtate de barplot for factors}
}
\value{
The function generates and saves barplots as PDF files in the project’s \verb{figures/} folder.
It also saves the processed data as CSV and RDS files in the corresponding \verb{output_data/} folders.
Additionally, the function outputs the plot object for further customization if needed.
}
\description{
This function generates barplots for microbial data at the genus level. It supports both
relative and absolute abundance data and can include facets based on available metadata factors.
The resulting plots can be saved as PDF files, and the underlying data can be exported as
CSV and RDS files.
}
\details{
\itemize{
\item Relative abundance plots show proportions of taxa in each sample, with taxa having a mean
relative abundance below 1\% grouped as "Other".
\item Absolute abundance plots use normalized cell equivalents (\code{norm_method = "fcm"} or \code{"qpcr"})
to display the number of cells per mL for each taxon.
\item Facets are added based on metadata factors present in the phyloseq object.
\item Taxa labels are styled to include genus and species names, if available.
}
}
\examples{
# Example usage
barplot(
  physeq = rarefied_genus_psmelt,
  ntaxa = 20,
  colorset = my_colors,
  norm_method = "fcm",
  sample_matrix = sample_metadata
)

}
