% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/decontam.R
\name{decontam}
\alias{decontam}
\title{Decontaminate a phyloseq object using specified methods}
\usage{
decontam(
  physeq = resolved_tree_physeq,
  decon_method = c("frequency", "prevalence", "both"),
  blank = TRUE,
  color_p1 = NULL
)
}
\arguments{
\item{physeq}{A \code{phyloseq} object containing the microbiome data.
This is the input object that the function processes.}

\item{decon_method}{A character string specifying the decontamination method.
this argument determines how the contamination removla will be applied.
The possible values are:
\itemize{
\item \code{frequency}: This method identifies contaminants based on their frequency of occurrece across samples.
Taxa that appear in very few samples (but with high abundace) are flagged as contaminants.
\item \code{prevalence}: This method identifies contamnants based on their prevalence in negative control sample (e.g., blank samples).
Taxa that are present in both negative control and true samples are flagged as contaminants.
\item \code{both}: This method combines both frequency and prevalence methods. It first applies the frequency
method, and then applies the prevalence method to identify additional contaminants. Taxa flagged by either
method are considered contaminants.
}
The default is \code{"both"}. However, if the \code{blank} parameter is set to \code{FALSE}, the method
is automatically changed to \code{"frequency"} for decontamination.}

\item{blank}{A logical value indicating whether blank samples were included in the dataset. Default is \code{TRUE}.
\itemize{
\item \code{TRUE}: Blank samples were included in the dataset. This allows the method \code{both} (frequency + prevalence)
to be used for decontamination by default, as blank samples are available for the prevalence based analysis.
\item \code{FALSE}: Blank sample were not included in the dataset. In this case, only the \code{frequency} method can be applied
for decontamination, as no blank samples are available for the prevalence method.
}}

\item{color_p1}{A string specifying the variable to use for coloring the read count plot.
Default is NULL, which uses "sample_type".}
}
\value{
A phyloseq object after decontamination. Contaminant taxa are removed based on
the chosen decontamination method.
}
\description{
This function performs contamination removal from a phyloseq object using either
frequency, prevalence, or both methods. It first checks for the presence of blank
samples and processes the contamination based on the specified method. The function
generates and saves plots showing the read counts and prevalence of contaminants.
}
\details{
The function checks if the "sample_type" column is present in the sample data of the
phyloseq object. It then proceeds to decontaminate the data using the specified
method. If 'both' is chosen as the decontamination method, both frequency and
prevalence methods are applied. The resulting decontaminated phyloseq object is returned,
and plots showing the read count and prevalence of contaminants are saved as PDF files.
}
\examples{
# Example usage:
decontam(physeq = my_phyloseq_data, decon_method = "both", blank = TRUE)

}
\seealso{
\code{\link{isContaminant}}, \code{\link{phyloseq}}, \code{\link{ggplot2}}
}
