% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/decontam.R
\name{decontam}
\alias{decontam}
\title{Decontaminate a Phyloseq Object Using Specified Methods}
\usage{
decontam(
  physeq = resolved_tree_physeq,
  decon_method = c("frequency", "prevalence", "both"),
  blank = TRUE
)
}
\arguments{
\item{physeq}{A \code{phyloseq} object containing the microbiome data.
This is the input object that the function processes.}

\item{decon_method}{A character string specifying the contamination removal method.
Possible values are:
\itemize{
\item \code{frequency}: Identifies contaminants by examining the distribution of sequence feature frequencies
as a function of the input DNA concentration.
\item \code{prevalence}: Identifies contaminants by comparing the prevalence (presence/absence across samples)
of sequence features in true samples versus negative controls (blanks).
\item \code{both}: Applies both frequency and prevalence methods sequentially.
Taxa flagged by either method are considered contaminants.
This option requires that blank samples are available.
}}

\item{blank}{A logical value indicating whether blank samples were included in the dataset.
\itemize{
\item \code{TRUE}: Blank samples are present, allowing the use of the \code{both} method.
\item \code{FALSE}: No blank samples are available; in this case, only the \code{frequency} method can be applied.
}}
}
\value{
A phyloseq object with contaminants removed. The decontaminated object is saved as an RDS file named
\verb{<project_name>_phyloseq_asv_level_decontam.rds} in the \code{output_data/rds_files/Before_cleaning_rds_files} directory.
}
\description{
This function removes contamination from a phyloseq object using the
\href{https://benjjneb.github.io/decontam/vignettes/decontam_intro.html}{\code{decontam}} package.
It supports the frequency, prevalence, or both methods for contaminant identification.
Note that the prevalence method relies on the presence of blank samples;
if blank samples are not available, only the frequency method can be performed.
}
\details{
The function uses the \code{decontam} package to remove contaminants based on the specified method.
If \code{both} is chosen, the function applies the frequency method first and then the prevalence method,
flagging any taxa identified by either method as contaminants.
In addition, diagnostic plots (showing read counts and contaminant prevalence) are generated and saved as PDF files.
}
\examples{
\dontrun{
# Example usage:
decontam(physeq = phyloseq_data, decon_method = "both", blank = TRUE)
}

}
