% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/group_tax.R
\name{group_tax}
\alias{group_tax}
\title{Aggregate Phyloseq Data by Taxonomic Level}
\usage{
group_tax(
  physeq = rarefied_asv_physeq,
  norm_method = NULL,
  taxrank = c("Phylum", "Class", "Order", "Family", "Tax_label")
)
}
\arguments{
\item{physeq}{A \code{phyloseq} object containing the microbiome data.
This is the input object that the function processes.}

\item{norm_method}{A string specifying the normalization method. Acceptable values:
\itemize{
\item \code{NULL}: Returns the \code{psdata_asv_copy_number_corrected} object without modifications.
\item \code{"fcm"}: Rarefied data using FCM normalized.
\item \code{"qpcr"}: Rarefied data using qPCR normalized.
}}

\item{taxrank}{A character vector indicating the taxonomic levels at which to group the data.}
}
\value{
The function saves multiple \code{phyloseq} objects as RDS files.
The aggregated objects are saved in the output directory \verb{output_data/rds_files/After_cleaning_rds_files/}.
}
\description{
This function aggregates ASV data at specified taxonomic levels (e.g., Phylum, Class, Order, Family, or Tax_label)
using the \code{tax_glom} function.
}
\details{
The function applies the \code{tax_glom} function to group ASVs at each specified taxonomic level. It creates
a dedicated folder for each taxonomic level under the output directory and saves the aggregated data as RDS files.
}
\examples{
\dontrun{
# Aggregate data using flow cytometry normalization
result <- group_tax(physeq = rarefied_asv_physeq, norm_method = "fcm")

# Aggregate data using qPCR normalization
result <- group_tax(physeq = rarefied_asv_physeq, norm_method = "qpcr")
}

}
