% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normalise_data.R
\name{normalise_data}
\alias{normalise_data}
\title{Normalize Phyloseq Data}
\usage{
normalise_data(
  physeq = without_mock_physeq,
  norm_method = NULL,
  copy_correction = TRUE
)
}
\arguments{
\item{physeq}{A \code{phyloseq} object containing the microbiome data.
This is the input object that the function processes.}

\item{norm_method}{A character string specifying the normalization method to use. Options are:
\itemize{
\item \code{"fcm"}: Normalize based on flow cytometry data, converting abundances to cell concentrations (cells/mL or gram sample).
\item \code{"qpcr"}: Normalize based on qPCR data, converting abundances to cell equivalents (cells/ml or gram sample).
\item \code{NULL}: Apply only copy number correction without additional normalization.
}}
}
\value{
A list containing one or more \code{phyloseq} objects:
\itemize{
\item \code{psdata_copy_number_corrected}: Phyloseq object after copy number correction.
\item \code{psdata_fcm_norm}: (if \code{norm_method = "fcm"}) Phyloseq object with FCM-normalized cell concentrations.
\item \code{psdata_qpcr_norm}: (if \code{norm_method = "qpcr"}) Phyloseq object with qPCR-normalized cell equivalents.
}
The normalized objects are also saved as \code{.rds} files in the specified output directory.
}
\description{
This function applies normalization to a \code{phyloseq} object, allowing the
conversion of data to absolute values based on either flow cytometry (FCM)
data or qPCR data. The funciton accounts for DNA and RNA normalization,
considering extra steps involved in prodessing RNA. Additionally, it
incorporates RasperGade16S correction to adjust ASV counts based on predicted copy
numbers. The normalized data can later be used for relative abundance calculations.
}
\details{
The function performs the following steps based on the chosen normalization method:
\enumerate{
\item \strong{Copy number correction (applied in all cases):}
\itemize{
\item Correct ASV abundances by dividing each count by the predicted copy number. The predicted 16S rRNA gene copy numbers are
based on the method described in
\href{https://dx.doi.org/10.1038/s43705-023-00266-0}{"Accounting for 16S rRNA copy number prediction uncertainty and its implications in bacterial diversity analyses"}.
\item This prediction enables the calculation of cell equivalents by adjusting for variablility in 16S rRNA gene copy numbers across taxa.
}
\item \strong{FCM Normalization (\code{norm_method = "fcm"}):}
\itemize{
\item Combine sample metadata (e.g., \code{cells_per_ml}) with corrected abundances.
\item Calculate absolute abundances as \code{cells_per_ml_sample} and adjust ASV abundances accordingly.
}
\item \strong{qPCR Normalization (\code{norm_method = "qpcr"}):}
\itemize{
\item Use qPCR-derived \code{sq_calc_mean} and predicted copy numbers to calculate absolute abundances.
\item Normalize abundances to represent cell equivalents, ensuring compatibility with downstream analysis.
}
}
\subsection{DNA vs RNA Normalization}{

The interpretation of normalized data depends on the type of nucleic acid (DNA or RNA):
\itemize{
\item \strong{DNA:} Normalized abundances typically represent \strong{cells per ml (or gram)}. This assumes one genome copy per cell.
\item \strong{RNA:} Normalized abundances often represent \strong{copies per cell equivalent per ml (or gram)}.
RNA reflects transcriptional activity and can vary widely depending on the condition of the cells.
}

When working with RNA data, ensure that metadata contains information about RNA extraction efficiency or other factors influencing RNA copy numbers.
}
}
\examples{
\dontrun{
# Apply only copy number correction
result <- normalize_data(physeq = physeq, norm_method = NULL)

# Normalize using flow cytometry (FCM) data
result <- normalize_data(physeq = physeq, norm_method = "fcm")

# Normalize using qPCR data
result <- normalize_data(physeq = physeq, norm_method = "qpcr")
}

}
\references{
Gao, Y., & Wu, M. (2023). Accounting for 16S rRNA copy number prediction
uncertainty and its implications in bacterial diversity analyses. ISME
communications, 3(1), 59. doi:\href{https://dx.doi.org/10.1038/s43705-023-00266-0}{10.1038/s43705-023-00266-0}
}
