% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/psdata_to_tibble.R
\name{psdata_to_tibble}
\alias{psdata_to_tibble}
\title{Convert Phyloseq Data to Tibble and Export by Taxonomic Level}
\usage{
psdata_to_tibble(
  physeq = rarefied_genus_physeq,
  norm_method = NULL,
  taxrank = c("Phylum", "Class", "Order", "Family", "Genus")
)
}
\arguments{
\item{physeq}{A \code{phyloseq} object containing the microbiome data.
This is the input object that the function processes.}

\item{norm_method}{A character string specifying the normalization method. Acceptable values are:
\itemize{
\item \code{NULL}: Use this option if no FCM or qPCR data is available, or if you wish to retain only relative abundances.
\item \code{"fcm"}: Use this option if the data have been normalized using flow cytometry (FCM).
\item \code{"qpcr"}: Use this option if the data have been normalized using quantitative PCR (qPCR).
}}

\item{taxrank}{A character vector indicating the taxonomic levels at which to group the data.}
}
\value{
The function saves multiple \code{phyloseq}-derived tibbles as RDS files and returns a list of tibbles:
\itemize{
\item \strong{If \code{norm_method} is \code{NULL}:}
A tibble of copy number窶田orrected counts is saved as
\code{<project_name>_psmelt_<tax>_level_copy_number_corrected_counts.rds} for each taxonomic level.

\if{html}{\out{<div class="sourceCode">}}\preformatted{\\item **If \code{norm_method = "fcm"}:**
      Two tibbles are saved for each taxonomic level:
      \\itemize\{
        \\item A tibble of copy number窶田orrected counts.
        \\item A tibble of FCM-normalized, rarefied counts saved as
              \code{<project_name>_psmelt_<tax>_level_fcm_normalised_cell_concentration_rarefied.rds}.
      \}

\\item **If \code{norm_method = "qpcr"}:**
      Two tibbles are saved for each taxonomic level:
      \\itemize\{
        \\item A tibble of copy number窶田orrected counts.
        \\item A tibble of qPCR-normalized, rarefied counts saved as
              \code{<project_name>_psmelt_<tax>_level_qpcr_normalised_cell_concentration_rarefied.rds}.
      \}
}\if{html}{\out{</div>}}

}
}
\description{
This function transforms a \code{phyloseq} object into a tibble for each specified taxonomic level
using the \code{psmelt} function from the \pkg{phyloseq} package.
}
\details{
The primary task of this function is to convert a \code{phyloseq} object into a tibble at each specified
taxonomic level using the \code{psmelt} function.
}
\examples{
\dontrun{
  # Export data without normalization (copy number窶田orrected counts only)
  result <- psdata_to_tibble(physeq = rarefied_genus_physeq)

  # Export data with flow cytometry normalization
  result <- psdata_to_tibble(physeq = rarefied_genus_physeq, norm_method = "fcm")

  # Export data with qPCR normalization
  result <- psdata_to_tibble(physeq = rarefied_genus_physeq, norm_method = "qpcr")
}

}
