% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/remove_mock.R
\name{remove_mock}
\alias{remove_mock}
\title{Remove Mock ASVs from Phyloseq Object}
\usage{
remove_mock(physeq = decontam_physeq, mock_genera = mock_genera, mock = TRUE)
}
\arguments{
\item{physeq}{A \code{phyloseq} object containing the microbiome data.
This is the input object that the function processes.}

\item{mock_genera}{A vector of genera representing the taxa that make up the mock community.
These are the taxa used to identify mock ASVs that can be removed from the \code{phyloseq} object.}

\item{mock}{A logical value (default \code{TRUE}) that determines whether or not to filter out the mock ASVs.
\itemize{
\item \code{TRUE}: Remove mock ASVs from the \code{phyloseq} object and retain only the samples for downstream analysis.
\item \code{FALSE}: Retain mock ASVs but filter the dataset to include only samples with \code{sample_type == "sample"}.
Use this option if no mock community is present in the dataset.
}}
}
\value{
A \code{phyloseq} object with filtered samples.
If \code{mock = TRUE}, mock ASVs are removed. If \code{mock = FALSE}, mock ASVs are retained, but the dataset includes only \code{sample_type == "sample"}.
The object is also saved as an RDS file for future use.
}
\description{
Remove Mock ASVs from Phyloseq Object
}
\details{
The function performs the following steps:
\itemize{
\item If \code{mock = FALSE}, the function filters the dataset to retain only samples with \code{sample_type == "sample"}, without removing mock ASVs.
This option is suitable for datasets where no mock community is included.
\item If \code{mock = TRUE}, the function:
\itemize{
\item Identifies mock ASVs based on the provided \code{mock_genera}.
\item Removes the mock ASVs from the dataset.
\item Retains only samples with \code{sample_type == "sample"}.
}
\item The filtered \code{phyloseq} object is saved as an RDS file with the name \code{project_name_phyloseq_asv_level_without_mock.rds}.
}
}
\examples{
\dontrun{
# Remove mock ASVs from the phyloseq object
physeq_no_mock <- remove_mock(physeq = physeq, mock_genera = c("Mock_Genus1", "Mock_Genus2"), mock = TRUE)

# Retain mock ASVs but filter to only 'sample' types (or when no mock community is present)
physeq_no_filter <- remove_mock(physeq = physeq, mock_genera = c("Mock_Genus1", "Mock_Genus2"), mock = FALSE)
}

}
