% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/remove_mock.R
\name{remove_mock}
\alias{remove_mock}
\title{Remove Mock Features from a Phyloseq Object}
\usage{
remove_mock(physeq = decontam_physeq, mock_genera = mock_genera, mock = TRUE)
}
\arguments{
\item{physeq}{A \code{phyloseq} object containing the microbiome data.
This is the input object that the function processes.}

\item{mock_genera}{A vector of genera representing the taxa that make up the mock community.
These taxa are used to identify mock features to be removed from the \code{phyloseq} object.}

\item{mock}{A logical value determining whether to filter out mock features.
\itemize{
\item \code{TRUE}: Remove mock features from the \code{phyloseq} object and retain only samples for downstream analysis.
\item \code{FALSE}: Retain mock features. Use this option if no mock community is present in the dataset.
}}
}
\value{
A filtered \code{phyloseq} object is returned and saved as an RDS file named
\code{project_name_phyloseq_asv_level_without_mock.rds} in the \verb{output_data/rds_files/Before_cleaning_rds_files/} directory.
}
\description{
This function removes mock features from a \code{phyloseq} object.
These mock features, which can appear in other samples due to cross-contamination,
are removed to minimize their impact on the analysis samples.
In addition, the function filters the dataset to retain only samples without controls.
Users can choose whether to remove the mock features by setting the \code{mock} parameter.
}
\details{
The function performs the following steps:
\itemize{
\item If \code{mock = FALSE}, the function filters the dataset to retain only samples without controls, leaving the mock features intact.
This option is suitable for datasets where no mock community is included.
\item If \code{mock = TRUE}, the function:
\itemize{
\item Identifies mock features based on the provided \code{mock_genera}.
\item Removes the mock features from the dataset.
\item Retains only samples without controls.
}
}
}
\examples{
\dontrun{
# Remove mock ASVs from the phyloseq object
physeq_no_mock <- remove_mock(physeq = physeq, mock_genera = c("Mock_Genus1", "Mock_Genus2"), mock = TRUE)

# Retain mock ASVs but filter to only samples (or when no mock community is present)
physeq_no_filter <- remove_mock(physeq = physeq, mock_genera = c("Mock_Genus1", "Mock_Genus2"), mock = FALSE)
}

}
