% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/resolve_tree.R
\name{resolve_tree}
\alias{resolve_tree}
\title{Resolve Multichotomies in Phyloseq Object's Tree}
\usage{
resolve_tree(physeq = cleaned_physeq)
}
\arguments{
\item{physeq}{A \code{phyloseq} object containing the microbiome data.
This is the input object that the function processes.}
}
\value{
A \code{phyloseq} object with a binary phylogenetic tree.
}
\description{
This function processes a \code{phyloseq} object to resolve multichotomies (non-binary nodes) in the phylogenetic tree
using the \code{ape} package. It ensures the tree is binary and saves the updated \code{phyloseq} object as an RDS file.
}
\details{
This function performs the following steps:
\itemize{
\item Checks if the \code{phyloseq} object contains a binary phylogenetic tree.
\item Resolves polychotomous nodes (if present) using \code{ape::multi2di()}.
\item Ensures that the tree is binary after resolution; raises an error if unresolved.
\item Merges the resolved tree back with the \code{otu_table}, \code{sample_data}, and \code{tax_table} in the \code{phyloseq} object.
\item Saves the updated \code{phyloseq} object with the resolved tree as an RDS file.
}
}
\examples{
\dontrun{
# Resolve tree and save the updated phyloseq object
resolved_physeq <- resolve_tree(physeq = cleaned_physeq)
}

}
