% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tax_clean.R
\name{tax_clean}
\alias{tax_clean}
\title{Clean Taxonomy Table in Phyloseq Object}
\usage{
tax_clean(physeq = physeq, tax_filter = TRUE)
}
\arguments{
\item{physeq}{A \code{phyloseq} object containing the microbiome data.
This is the input object that the function processes.}

\item{tax_filter}{Iindicating whether to apply additional
filtering to remove unwanted taxa (e.g., chloroplasts, mitochondria).
\itemize{
\item \code{TRUE}: Apply filtering to remove taxa such as Eukaryota, chloroplasts, mitochondria, and unclassified taxa.
\item \code{FALSE}: Skip filtering; the taxonomy table will only be cleaned but no taxa will be removed.
}}
}
\value{
A cleaned \code{phyloseq} object with ambiguous and unwanted taxa removed or replaced.
The cleaned \code{phyloseq} object is saved as an RDS file named \verb{<project_name>_phyloseq_cleaned.rds} in the \code{output_data/rds_files/Before_cleaning_rds_files} directory.
}
\description{
This function cleans and filters the taxonomy table within a \code{phyloseq} object.
It removes unclassified or ambiguous names, and replaces these and missing taxon names at genus level with placeholders derived from higher
taxonomic ranks. Optionally  this function filters specific taxa (e.g., Eukaryota, chloroplasts, mitochondria, and ASVs unclassified at Kingdom and Phylum levels).
The cleaned taxonomy enables taxonomic agglomeration (\code{tax_glom} form the package \code{phyloseq}) to genus level without merging unclassified taxa from diverse ancestry.
}
\details{
The function performs the following steps:
\itemize{
\item Cleans the taxonomy table:
\itemize{
\item Replaces ambiguous or placeholder taxa names (e.g., "uncultured organism") with \code{NA}.
\item Assigns missing taxon names based on their higher-level ranks (e.g., "Phylum of Kingdom").
\item Cleans the taxonomic ranks to remove overly nested placeholders.
\item Ensures consistent naming for \code{Kingdom} (e.g., replacing "d__Bacteria" with "Bacteria").
}
\item If \code{tax_filter = TRUE}, filters out unwanted taxa such as:
\itemize{
\item Chloroplasts (at \code{Class} or \code{Order} level).
\item Mitochondria (at \code{Family} level).
\item Unassigned taxa or taxa classified as \code{Eukaryota}.
}
\item Saves the cleaned \code{phyloseq} object as an RDS file.
\item Logs the number of ASVs (amplicon sequence variants) removed during cleaning.
}

The cleaned \code{phyloseq} object is returned for further analysis.
}
\examples{
\dontrun{
# Clean taxonomy table and apply filtering
physeq_cleaned <- tax_clean(physeq = physeq, tax_filter = TRUE)

# Clean taxonomy table without filtering
physeq_cleaned_no_filter <- tax_clean(physeq = physeq, tax_filter = FALSE)
}

}
