% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unify_metadata.R
\name{unify_metadata}
\alias{unify_metadata}
\title{Unify and Format Metadata}
\usage{
unify_metadata(projects)
}
\arguments{
\item{projects}{A character vector containing the names of the project (folders).}
}
\value{
A data frame containing the unified metadata. The data frame is also saved as a file in the project's \code{input_data} folder.
}
\description{
This function merges and formats metadata from various sources, including
QIIME metadata, experimental sample metadata, and qPCR or FCM data, to create a unified
metadata file for downstream analyses.
}
\details{
The function ensures that the metadata is unified and correctly formatted for further analysis by:
\itemize{
\item Reading and processing the \code{metadata_extra.tsv} file, which must contain at least:
\itemize{
\item \code{SampleID}: A unique identifier for each sample.
\item \code{sample_type}: Indicates whether the sample is a \code{sample}, \code{mock}, or \code{blank}.
\item \code{DNA_Concentration}: The DNA concentration (in ng/µl).
}
\item Optionally processing and integrating qPCR or FCM data:
\itemize{
\item If qPCR data is available, calculating the mean for duplicates and merging it with \code{metadata_extra}.
\item If FCM data is available, calculating the mean for duplicates and merging it with \code{metadata_extra}.
}
\item Reading the \code{metadata.tsv} file (QIIME metadata), ensuring it contains the \code{SampleID} column, and combining it with the processed \code{metadata_extra}.
\item Writing the final combined metadata to a file. Note that the output file is named by concatenating the project name with \verb{_metadata_formatted.tsv} and is saved in the project's \code{input_data} folder.
}

All metadata files (QIIME metadata, experimental sample metadata, and qPCR/FCM data) must include
the \code{SampleID} column for proper merging. This column serves as the key to align data from multiple sources.
}
\note{
This function requires that the folder structure has been set up (using the \code{create_folders} function) before running.
}
\examples{
\dontrun{
# Process and unify metadata for a project
unified_metadata <- unify_metadata(projects)
}

}
